tests/wpt/mozilla/tests/mozilla/bluetooth/getDescriptors/blacklisted-descriptors.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.blacklist);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [blacklist_test_service_uuid]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(blacklist_test_service_uuid))
    .then(service => service.getCharacteristic(blacklist_exclude_reads_characteristic_uuid))
    .then(characteristic => characteristic.getDescriptors())
    .then(descriptors => {
        assert_equals(descriptors.length, 1);
        assert_equals(descriptors[0].uuid, blacklist_exclude_reads_descriptor_uuid);
    });
}, 'The descriptors are blacklisted.');
</script>
